# lib/api-zod/src/generated/types/analysisList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResult } from "./analysisResult";

export interface AnalysisList {
  analyses: AnalysisResult[];
  total: number;
}