# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisResult {
  id: number;
  transactionSignature: string;
  userIntent: string;
  /** Whether the transaction matches the stated intent */
  intentMatch: boolean;
  /**
   * Risk score 0-100 (0 = safe, 100 = very risky)
   * @minimum 0
   * @maximum 100
   */
  riskScore: number;
  /** Non-technical plain-language explanation of the analysis */
  explanation: string;
  /** Detected transaction type */
  transactionType: string;
  /** Program IDs involved in the transaction */
  programs: string[];
  /** Specific warning flags raised */
  warnings: string[];
  analyzedAt: Date;
}