# lib/api-zod/src/generated/types/analyzeTransactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeTransactionBody {
  /**
   * Solana transaction signature or base64-encoded raw transaction
   * @minLength 1
   * @maxLength 2000
   */
  transaction: string;
  /**
   * Plain text description of what the user intends to do
   * @minLength 1
   * @maxLength 1000
   */
  userIntent: string;
  /**
   * Optional custom Solana RPC URL
   * @maxLength 200
   */
  rpcUrl?: string;
}